requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
sqlalchemy==2.0.25
python-dotenv==1.0.0
openai==1.7.2
